feat: mmkv logic

+++ chapter24/mmkv/src/sys.rs
@@ -1,4 +1,4 @@
-use std::ffi::{c_char, c_double, c_float, c_void};
+use std::ffi::{c_char, c_int, c_void};
 
 #[repr(C)]
 pub struct MMKV {
@@ -9,6 +9,6 @@ pub struct MMKV {
 extern "C" {
     pub fn get_mmkv_instance() -> *const MMKV;
     pub fn init_mmkv() -> c_void;
-    pub fn set_float(mmkv: *const MMKV, v: c_float, k: *const c_char) -> c_void;
-    pub fn get_float(mmkv: *const MMKV, k: *const c_char) -> c_float;
+    pub fn set_float(mmkv: *const MMKV, v: c_int, k: *const c_char) -> c_void;
+    pub fn get_float(mmkv: *const MMKV, k: *const c_char) -> c_int;
 }
